requirements.txt
discord-py-interactions
mysql-connector